Move types into primitives and update import paths
