+++ .gitignore
@@ -2,7 +2,7 @@
 
 # JetBrains iDEA
 .idea
-.vscode
 
 .DS_Store
 
+sandbox/
