+++ primitives/types/digest_item.go
