+++ primitives/types/hash.go
